+++ pkg/file/file.go
@@ -1,0 +1,66 @@
+package file
+
+import (
+	"io/ioutil"
+	"os"
+	"path/filepath"
+
+	"github.com/pkg/errors"
+)
+
+// Copies file by reading the file then writing atomically into the target directory
+func AtomicCopy(srcFilepath, targetDir, targetFilename string) error {
+	info, err := os.Stat(srcFilepath)
+	if err != nil {
+		return err
+	}
+
+	input, err := ioutil.ReadFile(srcFilepath)
+	if err != nil {
+		return err
+	}
+
+	return AtomicWrite(filepath.Join(targetDir, targetFilename), input, info.Mode())
+}
+
+// Write atomically by writing to a temporary file in the same directory then renaming
+func AtomicWrite(path string, data []byte, mode os.FileMode) (err error) {
+	tmpFile, err := ioutil.TempFile(filepath.Dir(path), filepath.Base(path)+".tmp.")
+	if err != nil {
+		return
+	}
+	defer func() {
+		if Exist(tmpFile.Name()) {
+			if rmErr := os.Remove(tmpFile.Name()); rmErr != nil {
+				if err != nil {
+					err = errors.Wrap(err, rmErr.Error())
+				} else {
+					err = rmErr
+				}
+			}
+		}
+	}()
+
+	if err = os.Chmod(tmpFile.Name(), mode); err != nil {
+		return
+	}
+
+	_, err = tmpFile.Write(data)
+	if err != nil {
+		if closeErr := tmpFile.Close(); closeErr != nil {
+			err = errors.Wrap(err, closeErr.Error())
+		}
+		return
+	}
+	if err = tmpFile.Close(); err != nil {
+		return
+	}
+
+	err = os.Rename(tmpFile.Name(), path)
+	return
+}
+
+func Exist(name string) bool {
+	_, err := os.Stat(name)
+	return err == nil
+}
